+++ backend/projects/serializers/__init__.py
@@ -1,0 +1,4 @@
+from .area import AreaSerializer
+from .project import ProjectSerializer, ProjectAreaSerializer
+from .installment import InstallmentSerializer
+from .overview import OverviewSerializer, ProjectSummarySerializer
