+++ backend/projects/serializers/area.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import Area
+
+
+class AreaSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Area
+        fields = ["id", "name", "projects"]
